% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brquasiControl.R, R/zzz_conventions.R
\name{brquasiControl}
\alias{brquasiControl}
\alias{quasi_control}
\alias{brquasi_control}
\title{Auxiliary function for [`glm()`] fitting using the
[`brquasiFit()`] method.}
\usage{
brquasiControl(
  epsilon = 1e-04,
  maxit = 500,
  trace = FALSE,
  type = c("iRBM", "M", "eRBM", "MPQL_trace"),
  slowit = 1,
  response_adjustment = 0,
  max_step_factor = 12,
  only_beta = TRUE,
  lambda = 1e-10,
  disp_factor = c("n-p", "n")
)

brquasi_control(
  epsilon = 1e-04,
  maxit = 500,
  trace = FALSE,
  type = c("iRBM", "M", "eRBM", "MPQL_trace"),
  slowit = 1,
  response_adjustment = 0,
  max_step_factor = 12,
  only_beta = TRUE,
  lambda = 1e-10,
  disp_factor = c("n-p", "n")
)
}
\arguments{
\item{epsilon}{positive convergence tolerance epsilon. Default is
`1e-04`.}

\item{maxit}{integer giving the maximal number of iterations
allowed. Default is `500`.}

\item{trace}{logical indicating if output should be produced for
each iteration. Default is `FALSE`.}

\item{type}{the type of fitting method to be used. The options are
`"M"` (standard M-estimation based on quasi-likelihoods),
`"iRBM"` (implicit reduced-bias M-estiamtion: empirically
adjusted quasi likelihood equations for mean-bias reduction;
default), `eRBM` (explicit reduced-bias M-estimation:
correction of asymptotic mean bias using empirical bias
estimates). `"MPQL_trace"` (maximum penalized quasi likelihood
estimation; see Details).}

\item{slowit}{a positive real used as a multiplier for the
stepsize. The smaller it is the smaller the steps are. Default
is `1`.}

\item{response_adjustment}{a (small) positive constant or a vector
of such. Default is `NULL`. See Details.}

\item{max_step_factor}{the maximum number of step halving steps to
consider. Default is `12`.}

\item{only_beta}{Should RBM estimation be used for improving
estimation of the regression coefficients only? Default is
`TRUE`.}

\item{lambda}{a ridge adjustment to be added to the diagonal of the
jacobian before inverting it for the computation of the step
size in the quasi-Fisher iteration. Default is `1e-10`. See
Details.}

\item{disp_factor}{factor by which to divide the sum of pearson
residuals when estimating the dispersion. Defauls is `"n-p"`
which corresponds to "number of observations" minus
"number of parameters". See Details.}
}
\description{
Typically only used internally by `brquasiFit()`, but may be used
to construct a `control` argument.
}
\details{
TO ADD: details on iRBM, eRBM, disp_factor
}
