% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brquasiFit.R, R/zzz_conventions.R
\name{brquasiFit}
\alias{brquasiFit}
\alias{brquasi_fit}
\title{Fitting function for [`glm()`] for reduced-bias
estimation and inference using quasi likelihoods}
\usage{
brquasiFit(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)

brquasi_fit(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\arguments{
\item{x}{`x` is a design matrix of dimension `n * p`.}

\item{y}{`y` is a vector of observations of length `n`.}

\item{weights}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.}

\item{start}{starting values for the parameters in the linear
predictor. If `NULL` (default) then the maximum likelihood
estimates are calculated and used as starting values.}

\item{etastart}{applied only when start is not
`NULL`. Starting values for the linear predictor to be
passed to `glm.fit` when computing starting values
using maximum likelihood.}

\item{mustart}{applied only when start is not `NULL`. Starting
values for the vector of means to be passed to
`glm.fit` when computing starting values using
maximum likelihood.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one is
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}

\item{family}{any one of the [`quasi()`] families.}

\item{control}{a list of parameters controlling the fitting
process. See [`brquasiControl()`] for details.}

\item{intercept}{logical. Should an intercept be included in the
    \emph{null} model?}

\item{singular.ok}{logical. If `FALSE`, a singular model is an
error.}
}
\description{
`brquasiFit` is a fitting method for [`glm()`] for
mean-bias reduction in quasi-likelihood estimation, using the
implicit and explicit reduced-bias M-estimators in Kosmidis &
Lunardon (2020).
}
\details{
Estimation is performed using a quasi Fisher scoring iteration
(similar to what is described, for example, in
`vignette("iteration", "brglm2")`, which, in the case of
mean-bias reduction, resembles an iterative correction of the
asymptotic bias of the Fisher scoring iterates.
}
\seealso{
[`glm.fit()`] [`glm()`] [`quasi()`]
}
\author{
Ioannis Kosmidis [aut, cre] \email{ioannis.kosmidis@warwick.ac.uk}
}
